Move `WithBase` to common sources
